requirements/base.txt
# python modules
gunicorn>=19.6,<19.7 # server
psycopg2>=2.6.2,<2.7 # postgres bindings
pillow>=4.0,<4.1 # image manipulation

Django>=1.8,<1.9

django-contrib-comments>=1.7.3,<1.8
django-localflavor>=1.4,<1.5

django-compressor>=2.1,<2.2
django-supervisor>=0.4.0,<0.5
django-filebrowser>=3.7,<3.8
django-grappelli>=2.9,<2.10
django-tinymce>=2.4,<2.5
